+++ src/middleware.rs
@@ -1,0 +1,5 @@
+//! Experimental middleware.
+//!
+//! Analogous to the `middleware` module in Actix Web.
+
+pub use crate::middleware_from_fn::{from_fn, MiddlewareFn, Next};